requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
boto3==1.21.24
botocore==1.24.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Deprecated==1.2.13
frozenlist==1.3.0
future==0.18.2
idna==3.3
jmespath==1.0.0
jwcrypto==1.0
line-bot-sdk==2.1.0
multidict==6.0.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
